move electron_paths.h to shell/common
