+++ filenames.gni
@@ -174,7 +174,6 @@ filenames = {
     "shell/browser/electron_javascript_dialog_manager.h",
     "shell/browser/electron_navigation_throttle.cc",
     "shell/browser/electron_navigation_throttle.h",
-    "shell/browser/electron_paths.h",
     "shell/browser/electron_permission_manager.cc",
     "shell/browser/electron_permission_manager.h",
     "shell/browser/electron_quota_permission_context.cc",
@@ -474,6 +473,7 @@ filenames = {
     "shell/common/electron_command_line.h",
     "shell/common/electron_constants.cc",
     "shell/common/electron_constants.h",
+    "shell/common/electron_paths.h",
     "shell/common/gin_converters/accelerator_converter.cc",
     "shell/common/gin_converters/accelerator_converter.h",
     "shell/common/gin_converters/blink_converter.cc",
