+++ shell/browser/ui/devtools_manager_delegate.cc
@@ -27,7 +27,7 @@
 #include "net/base/net_errors.h"
 #include "net/socket/stream_socket.h"
 #include "net/socket/tcp_server_socket.h"
-#include "shell/browser/electron_paths.h"
+#include "shell/common/electron_paths.h"
 #include "ui/base/resource/resource_bundle.h"
 
 namespace electron {
